Find and read refs in a repo.
Added functions for find refs with a certain path in a repo and reading
them.
Also added functions for finding all refs in a repo.
Basic unit tests for these newly defined functions.
Repo structs now store the gitDir, not the repoDir (i.e. the parent of
the git dir) since the worktree and git dir may be different.
Added FindRef and GetAllRef on repo structs which are wrappers on the
functions defined in refs.go.