Added functions for creating an index.
Manually add index entries based on a file's property.

Add extension data (which are ignored for now). Some extensions that
change the format of the index file will break this program.

Parse an existing index created by git, as well as convert the internal
stucts used to represent an index back into the format git understands.